requirements.txt
streamlit>=1.32.0


pymupdf>=1.23.0


easyocr>=1.7.1
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0


langchain>=1.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0


faiss-cpu>=1.7.4


sentence-transformers>=2.6.0
transformers>=4.38.0
huggingface-hub>=0.20.0


openai>=1.14.0


tqdm>=4.66.0